Bug 23206: Batch patron modification from reports which return cardnumber
This patch adds an option to submit patron card numbers from report
results to the batch patron modification tool.

As with a report which returns biblionumbers or itemnumbers, a button
appears at the top of report results which triggers a menu of batch
operation options.

To test, apply the patch and create a report which includes patron card
number in the results.

 - Run the report and confirm that under the "Batch operations" button
   a batch patron modification menu item appears.
   - Confirm that selecting this option sends the correct patrons to the
     batch patron modification tool.
 - Test also with a report which returns both patron cardnumber and
   biblionumber or itemnumber.

Signed-off-by: Maryse Simard <maryse.simard@inlibro.com>
Signed-off-by: George <george@nekls.org>
Signed-off-by: Josef Moravec <josef.moravec@gmail.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>