Bug 22951: Markup error in OPAC holds template
This patch makes minor markup changes to the OPAC holds template in
order to fix validation errors.

To test, apply the patch and go through the process of placing a
hold in the OPAC. Test various aspects of the process, including cases
where the hold is blocked and an error message is displayed (for
instance, when there are no items to be placed on hold).

The page should work correctly and validating the source should
return no relevant errors.

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>
Signed-off-by: Frederik Chenier <frederik.chenier@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
(cherry picked from commit e0991ae2358224a15b6880c96b8472a3cd814b1b)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>