Bug 23143: [18.11] Filter paid transactions not working
This patch corrects the DataTables filter for hiding paid transactions
from the table of patron account transactions. The wrong column index
was specified.

To test, apply the patch and view the "Account" tab of a patron with
multiple transactions, at least one of them paid.

Confirm that the "Filter paid transactions" link works as expected.

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>