Bug 23039: Hold found modal on checkin screen ( circulation.pl ) obscures Check in message info
This patch modifies the checkin process so that any errors or messages
related to the current checkin are displayed in any hold/transit modal
which is displayed. This prevents any information from being hidden by
the modal.

This patch also adds an "information" button to the checkout form which
the user can click to redisplay the last modal which was dismissed
(unless confirming the modal reloads the page).

To test, apply the patch and rebuild the staff client CSS
(https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client).

 - Test the checkin process with various types of items:
   - Item not checked out
   - Item not checked out, on hold
   - Item not checked out, needs to be transferred
   - Item is marked lost, needs to be transferred
   - Item checked out, patron has waiting holds (with
     WaitingNotifyAtCheckin enabled)
   - Item checked out, patron has outstanding fines (with
     FineNotifyAtCheckin enabled)

 - Test that the modal redisplay button works in cases where the modal
   is dismissed without triggering a page reload:
   - Hold confirmation -> Ignore
   - Transfer confirmation -> OK

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Bouzid Fergani <bouzid.fergani@inlibro.com>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>