Bug 21518: Material type "three-dimensional artifact" displays as "visual material"
This patch makes a few corrections to the staff client MARC21 XSLT so
that the correct material type icons are displayed for pictures and
realia. The more-specific material type handling has been adapted from
the corresponding XSL files in the OPAC.

To test you must have default XSLT enabled and the DisplayIconsXSLT and
DisplayOPACiconsXSLT system preferences enabled.

 - Apply the patch and locate records with "type of record" specified in
   the leader as 'g', 'k', and 'r'.

   A sample MARC file will be attached to the bug report with one of
   each type of record. Import the records and search your catalog for
   "Bug_21518."

 - View the records in search results and their detail pages, both in
   the OPAC and the staff client. In all cases the correct material type
   icon should be displayed.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>