Bug 23441: Export should not include the 'actions' column in Z3950 results.
This patch adds the "noExport" class to the table header for the Z39.50
search results "Actions" column. This class prevents the column from
being included in DataTables export operations.

To test, apply the patch and perform a Z39.50 search from Cataloging. On
the search results page, click the "Export or print" button. Choose one
of the export options and verify that in the resulting data there is no
"Actions" column.

Signed-off-by: Maryse Simard <maryse.simard@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>