Bug 17057: Remove event attributes from holds template
My orginal patch for this bug contained a lot of changes made obsolete
by Bug 19469, which created holds_table.inc.

This new version makes one minor change only to request.tt, adding a
class name and removing an onclick attribute.

To test, apply the patch and place an item-level hold on a title.

Start the process of placing another hold on that item. On the "place
hold" screen, in the table of items, you should see a "Cancel hold" link
for the item you previously placed on hold.

 - Click the "Cancel hold" link
   - Clicking "Cancel" in the confirmation dialog should cancel the
     operation.
   - Clicking "OK" should cancel the hold.

Signed-off-by: Maryse Simard <maryse.simard@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>