Bug 17057: (follow-up) Replace onsubmit attributes
This patch removes the "onsubmit" attribute from the place hold form.

To test, apply the patch and test placing a hold:

Begin the process of placing a hold on a single title. On the
place hold screen, uncheck the "Hold next available item" checkbox and
click the "Place hold" button.

You should be get an alert saying "Please select an item to place a
hold." Select an item and click the "Place hold" button again. The hold
should be placed.

This patch affects the form validation for placing multiple holds, but I
don't think the JavaScript ever has a chance to run: The template blocks
the user from placing multiple holds if one of the titles has no items.

Signed-off-by: Maryse Simard <maryse.simard@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>