Bug 23445: (RM follow-up) Add warning for unexpected lengthunit
Discussions on this bug highlighted worries about behind the scenes
fallbacks to 'days' as a lengthunit should the lengthunit be set to an
unrecognised value. This patch adds a warning to the about page for any
such occurrences so librarians may be made aware of and resolve the issue.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
(cherry picked from commit 1aae63606caeb389a40bd644b1f9e30917325573)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>